# GS/HW: Detect double buffers when Z not offset during RT in RT