@JordanTheToaster 2025-05-01
Improves performance in Stolen and 007 FRWL. Stolen at 6x goes from 28 fps to 80 fps and 007 FRWL at 12x goes from 59 fps to 108 fps.
![image](https://github.com/user-attachments/assets/999b4c73-2c41-4677-95f8-115fbfc6d90b)
![image](https://github.com/user-attachments/assets/b99d8f8a-f6d8-4bd2-a052-9381a6febaf5)